automata_exp v0.1.0 AutomataExp.DFA

Deterministic Finite Automata (DFA)

This module defines deterministic finite automata (DFA), which are usually used to recognize regular languages.

Read the documentation for new/5 to learn how to define an automaton. The documentation for run/4 shows how to make the automaton do some work. Also, the following demo functions show typical usages of DFAs:

AutomataExp.Demo.check_parity/1

AutomataExp.Demo.recognize_identifier/1

AutomataExp.Demo.recognize_integer/1

Types

hook_fn()

hook_fn() :: (state(), term(), integer() -> term())

result()

result() :: {:accepted | :rejected, term(), charlist()}

state()

state() :: term()

t()

t() :: %AutomataExp.DFA{
  accepted_states: MapSet.t(state()),
  initial_state: state(),
  states: MapSet.t(state()),
  symbols: MapSet.t(integer()),
  transition_fn: transition_fn()
}

transition_fn()

transition_fn() :: (state(), integer() -> state() | nil)

Functions

new(q, s, d, q0, f)

new([state()], charlist(), transition_fn(), state(), [state()]) :: t()

Creates a new DFA struct.

Use this function to define a deterministic finite automaton. Creating a struct directly using %DFA{} is not recommended.

Arguments

q: A set of states a new automaton will have (Q).

s: A set of terminal symbols accepted by this automaton (Σ).

d: A transition function (δ). Read further for more information.

q0: A starting state (q₀ ∈ Q).

f: A set of accepting states (F ⊆ Q).

Arguments q, s, and f accept lists, instead of MapSets, to minize the verbosity of your code. Those lists are converted into MapSets by this function. You can use any Elixir term for each state, but atoms are suitable for most of the cases. However, s must be a list of characters (charlist), such as '01', 'abcdefgh', etc.

Defining a Transition Function (δ)

The following rules must be met for each transition function (δ):

The function must accept two arguments. The first one is the current state of the automaton, and the second one is the current input character.

Each function clause must return the next state of the automaton.

An extra function clause (_, _ -> nil) shall follow the last defined clause. This clause takes a role of ∅ (an empty set).

Here are some examples of transition functions:

fn
  :even, ?0 -> :even
  :even, ?1 -> :odd
  :odd, ?0 -> :odd
  :odd, ?1 -> :even
  _, _ -> nil
end

# You can also use guards to accept a character in a range.
fn
  :first, x when x in ?a..?z -> :rest
  :rest, x when x in ?a..?z or x in ?0..?9 -> :rest
  _, _ -> nil
end

Most of the time, using guards to specify a range of accepted characters is tedious and requires a lot of keystrokes. You can use AutomataExp.Macros.dfa_fn/1 macro to alleviate this problem.

# Make sure you import and require AutomataExp.Macros module.

dfa_fn do
  # Single character
  :q0, ?a -> :q1

  # Single range
  :q1, ?w..?z -> :q2

  # A crazy combination
  :q2, [?p, ?q, '+-*/', ?0..?9] -> :q3
end

The code above is roughly transformed into this:

fn
  # Single character
  :q0, ?a -> :q1

  # Single range
  :q1, x when x in ?w..?z -> :q2

  # A crazy combination
  :q2, x when x == ?p or x == ?q or x in '+-*/' or x in ?0..?9 -> :q3

  # This clause is automatically appended.
  _, _ -> nil
end

run(dfa, chars, acc \\ nil, hook_fn \\ fn _, a, _ -> a end)

run(t(), charlist(), term(), hook_fn()) :: result()

Runs the given DFA.

This function makes the given DFA process the input charlist chars. The automaton will read a character from chars and move around the states according to the transition function.

If you additionally specify an accumulator value acc and a hook function hook_fn, The automaton will call the hook_fn every time it does a transition. hook_fn must take three arguments: the new state of the automaton, the current accumulator, and a character the automaton just read. And it must return the new value of the accumulator. For example, let’s say that a DFA in state :foo moved to state :bar by reading a character a, and the current accumulator value is 100. Then the hook function will be called with arguments :bar, 100, and ?a.

Return Value

Values returned by this function are in the form of {status, acc, rest}. status is one of :accepted or :rejected, and it indicates whether the automaton has accepted or rejected the input charlist. acc is the final value of the accumulator, which usually is meaningful only if status is :accepted. And finally, rest is a part of the input charlist which has not been read by the automaton before it halted. rest must be an empty list if status is :accepted, obviously.