automata_exp v0.1.0 AutomataExp.Demo

Provides some demo functions of automata.

Source codes are available on GitHub.

Functions

check_parity(chars)

check_parity(charlist()) :: any()

Checks if the given charlist is a string of 0s and 1s where the number of 1s is even.

If the number of 1s is even, the automaton accepts the input.

Examples

iex> check_parity('0000')
{:accepted, nil, []}

iex> check_parity('0101')
{:accepted, nil, []}

iex> check_parity('1101')
{:rejected, nil, []}

recognize_identifier(chars)

recognize_identifier(charlist()) :: any()

Recognizes an identifier name.

In order for the automaton to accept the input, it must follow these rules:

The first character must be a lowercase letter (a-z) or an underscore (_).

The rest of characters can be a mixture of lowercase letters, an underscore, and decimal digits (0-9).

Examples

iex> recognize_identifier('a')
{:accepted, "a", []}

iex> recognize_identifier('_')
{:accepted, "_", []}

iex> recognize_identifier('to_string')
{:accepted, "to_string", []}

iex> recognize_identifier('__callback')
{:accepted, "__callback", []}

iex> recognize_identifier('99_bottles_of_beer')
{:rejected, "", '99_bottles_of_beer'}

recognize_integer(chars)

recognize_integer(charlist()) :: any()

Recognizes an integer literal and returns its value.

See the examples below to see which formats are accepted.

Examples

# Zero
iex> recognize_integer('0')
{:accepted, 0, []}

# Decimal
iex> recognize_integer('1234')
{:accepted, 1234, []}

# Octal
iex> recognize_integer('0755')
{:accepted, 493, []}

# Invalid octal
iex> recognize_integer('0493')
{:rejected, 4, '93'}

# Prefix for hexadecimal integer (invalid)
iex> recognize_integer('0x')
{:rejected, 0, []}

# Hexadecimal (in lowercase)
iex> recognize_integer('0xac00')
{:accepted, 44032, []}

# Hexadecimal (in mixed-case)
iex> recognize_integer('0XD7a3')
{:accepted, 55203, []}

# Invalid hexadecimal
iex> recognize_integer('0x12Z9')
{:rejected, 18, 'Z9'}